Moved code for the `/supplier/{}` route to a new directory: `/app/main/views`. Included `row2dict` lamda snippet that I found on stackoverflow.
